break up ref_resolve into modules
