+++ packages/doenetml-worker/lib-doenetml-core/src/dast/mod.rs
@@ -15,5 +15,6 @@ pub mod dast_structure;
 pub mod flat_dast;
 pub mod ref_expand;
 pub mod ref_resolve;
+pub mod root_names;
 
 pub use dast_structure::*;
